requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
cryptography
pyjwt
paho-mqtt
psutil
